refactor(settings): move the org API keys table to its own file
The extracted table is a full component with its own hooks and table UI, so
it belongs in its own file under the repo's one-component-per-file rule. Its
private helpers move with it.

Generated-By: PostHog Desktop
Task-Id: 8b1fb267-a3d3-4187-9003-097320ffb077
